update branch with develop, now uses styles and themes from the common module
