+++ app/src/main/assets/paperwork.json
@@ -1,1 +1,1 @@
-{"buildTime":"2021-07-02 15:28","gitSha":"072b5eed8"}
+{"buildTime":"2021-07-02 15:29","gitSha":"072b5eed8"}
