--- commons/src/main/java/org/dhis2/commons/customdialogs/CalendarPickerPresenter.kt
@@ -1,6 +1,0 @@
-package org.dhis2.commons.customdialogs
-
-interface CalendarPickerPresenter {
-    fun isDatePickerStyle(): Boolean
-    fun setPickerStyle(isDatePicker: Boolean)
-}
