+++ commons/src/main/java/org/dhis2/commons/dialogs/calendarpicker/CalendarPickerRepository.kt
@@ -1,0 +1,6 @@
+package org.dhis2.commons.dialogs.calendarpicker
+
+interface CalendarPickerRepository {
+    fun isDatePickerStyle(): Boolean
+    fun setPickerStyle(isDatePicker: Boolean)
+}
