+++ commons/src/main/java/org/dhis2/commons/dialogs/calendarpicker/di/CalendarPickerModule.kt
@@ -1,0 +1,18 @@
+package org.dhis2.commons.dialogs.calendarpicker.di
+
+import dagger.Module
+import dagger.Provides
+import org.dhis2.commons.dialogs.calendarpicker.CalendarPickerRepository
+import org.dhis2.commons.dialogs.calendarpicker.CalendarPickerRepositoryImpl
+import org.dhis2.commons.prefs.PreferenceProvider
+
+@Module
+class CalendarPickerModule {
+
+    @Provides
+    fun providesCalendarPickerPresenter(
+        preferences: PreferenceProvider
+    ): CalendarPickerRepository {
+        return CalendarPickerRepositoryImpl(preferences)
+    }
+}
